switched to using non-f math funcs

+++ src/panner.cpp
@@ -298,7 +298,7 @@ EqualPowerStereoPanner::update ()
 	*/
 
 	const float pan_law_attenuation = -3.0f;
-	const float scale = 2.0f - 4.0f * powf (10.0f,pan_law_attenuation/20.0f);
+	const float scale = 2.0f - 4.0f * pow (10.0, pan_law_attenuation/20.0);
 
 	/* x == 0 => hard left
 	   x == 1 => hard right
